attempt at fixing circular imports again
